Move allegedb's window functions to their own module
This is mostly to get rid of an error in the Sphinx build but might
prevent some import errors in the future.
